+++ backend/src/main/java/com/debede/backend/repository/Carrito_ProductoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Carrito_ProductoRepository extends JpaRepository<Carrito_Producto, Long> {
+public interface Carrito_ProductoRepository extends JpaRepository<Carrito_Producto, Integer> {
 
 }
